pyproject.toml
[tool.poetry]
name = "neighborhub"
version = "0.1.0"
description = ""
authors = ["bilalimamoglu <xfyzal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.30.0"
streamlit-calendar = "^1.1.0"
gspread = "^5.12.4"
oauth2client = "^4.1.3"
setuptools = "^69.0.3"
pydeck = "^0.8.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
